+++ @navikt/core/react/src/form/combobox/Combobox.tsx
@@ -7,11 +7,11 @@ import ComboboxWrapper from "./ComboboxWrapper";
 import ClearButton from "./parts/buttons/ClearButton";
 import ToggleListButton from "./parts/buttons/ToggleListButton";
 import FilteredOptions from "./parts/filtered-options/FilteredOptions";
-import { useFilteredOptionsContext } from "./parts/filtered-options/filteredOptionsContext";
+import { useFilteredOptionsContext } from "./parts/filtered-options/FilteredOptions.context";
 import Input from "./parts/input/Input";
 import { useInputContext } from "./parts/input/Input.context";
 import SelectedOptions from "./parts/selected-options/SelectedOptions";
-import { useSelectedOptionsContext } from "./parts/selected-options/selectedOptionsContext";
+import { useSelectedOptionsContext } from "./parts/selected-options/SelectedOptions.context";
 
 export const Combobox = forwardRef<HTMLInputElement, ComboboxProps>(
   (props, ref) => {
