+++ @navikt/core/react/src/form/combobox/parts/filtered-options/FilteredOptions.context.tsx
@@ -12,7 +12,7 @@ import { ComboboxBaseProps, ComboboxOption } from "../../Combobox.types";
 import { useCustomOptionsContext } from "../../customOptionsContext";
 import { toComboboxOption } from "../../util/combobox-utils";
 import { useInputContext } from "../input/Input.context";
-import { useSelectedOptionsContext } from "../selected-options/selectedOptionsContext";
+import { useSelectedOptionsContext } from "../selected-options/SelectedOptions.context";
 import filteredOptionsUtils from "./filtered-options-util";
 import useVirtualFocus, { VirtualFocusType } from "./useVirtualFocus";
